requirements.txt
streamlit==1.30.0
torch==2.8.0
transformers==4.34.0
openai-whisper==20250625
ffmpeg-python==0.2.0
numpy==1.26.4
pandas==2.3.2
opencv-python-headless==4.8.1.78
tqdm==4.67.1
rich==13.9.4
moviepy==1.0.3
